% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add}
\alias{add}
\title{Add to existing values}
\usage{
add(x, add = 1)
}
\arguments{
\item{x}{variable to modify}

\item{minus}{constant to subtract from original value; default = 1}
}
\value{
adjusted numeric vector
}
\description{
Adds constant to values
}
