% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_prefix}
\alias{add_prefix}
\title{Add prefix to variable name}
\usage{
add_prefix(x, prefix)
}
\arguments{
\item{x}{variable to be renamed}

\item{prefix}{character string to be added}
}
\value{
renamed variable
}
\description{
Renames variable by adding character string before existing variable name
}
