% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_qualtrics}
\alias{clean_qualtrics}
\title{Do a preliminary cleaning of Qualtrics data}
\usage{
clean_qualtrics(data)
}
\arguments{
\item{data}{dataframe to be cleaned}
}
\value{
dataframe
}
\description{
Removes non-value rows and changes variable names to lowercase
}
