% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dichotomize}
\alias{dichotomize}
\title{Dichotomize values}
\usage{
dichotomize(x, cutoff, group0 = 0, group1 = 1)
}
\arguments{
\item{x}{variable to be dichotomized}

\item{cutoff}{value used to split groups; <= cutoff}

\item{group0}{value assigned to members of first group; default = 0}

\item{group1}{value assigned to members of second group; default = 1}
}
\value{
binary vector
}
\description{
Classifies existing values into two new categories or groups
}
