% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{histograms}
\alias{histograms}
\title{Plot multiple histograms}
\usage{
histograms(data, binwidth = 1, ...)
}
\arguments{
\item{data}{data frame}

\item{binwidth}{binwidth in histogram; default = 1}

\item{...}{other arguments that are passed to the ggplot() function}
}
\value{
plot in R environment
}
\description{
Plots histogram for each vector in data frame
}
