% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{minus}
\alias{minus}
\title{Subtract from existing values}
\usage{
minus(x, minus = 1)
}
\arguments{
\item{x}{variable to modify}

\item{minus}{constant to subtract from original value; default = 1}
}
\value{
adjusted numeric vector
}
\description{
Subtracts constant from values
}
