% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{qqplots}
\alias{qqplots}
\title{Plot multiple quantile-quantile plots}
\usage{
qqplots(data, ...)
}
\arguments{
\item{data}{data frame}

\item{...}{other arguments that are passed to the ggplot() function}
}
\value{
plot in R environment
}
\description{
Plots q-q plot for each vector in data frame
}
