% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{rci_dec}
\alias{rci_dec}
\title{Calculate reliable change index for symptoms}
\usage{
rci_dec(x_pre, x_post)
}
\arguments{
\item{x_pre}{vector for scores at pretest}

\item{x_post}{vector for scores at posttest}
}
\value{
categorical vector indicating RCI status
}
\description{
Calculates reliable change index based on criterion defined in Jacobson & Traux (1991) and assumes decrease is improvement
}
