% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recode}
\alias{recode}
\title{Recode variables}
\usage{
recode(x, value = 1, final = 99)
}
\arguments{
\item{x}{variable to recode}

\item{value}{value to recode; default = 1}

\item{final}{desired value; default = 99}
}
\value{
recoded numeric vector
}
\description{
Recodes existing variables to new constant
}
