% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_prefix}
\alias{remove_prefix}
\title{Remove prefix from variable name}
\usage{
remove_prefix(x, prefix)
}
\arguments{
\item{x}{variable to be renamed}

\item{prefix}{character string to be removed}
}
\value{
renamed variable
}
\description{
Renames variable by removing character string from existing variable name
}
