% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reverse}
\alias{reverse}
\title{Recode reverse scored values}
\usage{
reverse(x, highest = 6)
}
\arguments{
\item{x}{variable to be reverse coded}

\item{highest}{the number from which to subtract the observed value; default = 6}
}
\value{
correctly scored numeric vector
}
\description{
Recodes variables that are reverse scored on Likert-type scales
}
