% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{scatterplots}
\alias{scatterplots}
\title{Plot multiple scatterplots}
\usage{
scatterplots(data, y, x, fit = "lm", ...)
}
\arguments{
\item{data}{data frame}

\item{y}{dependent variable}

\item{x}{subsetted data frame of independent variables}

\item{fit}{specify modeling function for best fitting line, default = "lm"}

\item{...}{other arguments that are passed to the ggplot() function}
}
\value{
plot in R environment
}
\description{
Plots scatterplot for DV and each specified IV in data frame
}
