% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_numeric}
\alias{to_numeric}
\title{Convert factors to numeric vectors}
\usage{
to_numeric(x)
}
\arguments{
\item{x}{variable to be converted}
}
\value{
numeric vector
}
\description{
Converts factors to numeric vectors without changing original values
}
