% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{remove_prefix}
\alias{remove_prefix}
\title{Remove prefix from variable names}
\usage{
remove_prefix(data, prefix)
}
\arguments{
\item{data}{data where variables are located}

\item{prefix}{character string to be removed}
}
\value{
renamed variable
}
\description{
Renames variables by removing character string from existing variable names
}
